feat(core): support decimal multi-asset trade flows

+++ core/src/commands/order.rs
@@ -6,7 +6,7 @@ use uuid::Uuid;
 
 pub fn create_stop(
     trading_vehicle_id: Uuid,
-    quantity: i64,
+    quantity: Decimal,
     price: Decimal,
     currency: &Currency,
     category: &TradeCategory,
@@ -25,7 +25,7 @@ pub fn create_stop(
 
 pub fn create_stop_with_category(
     trading_vehicle_id: Uuid,
-    quantity: i64,
+    quantity: Decimal,
     price: Decimal,
     currency: &Currency,
     category: &TradeCategory,
@@ -47,7 +47,7 @@ pub fn create_stop_with_category(
 
 pub fn create_entry(
     trading_vehicle_id: Uuid,
-    quantity: i64,
+    quantity: Decimal,
     price: Decimal,
     currency: &Currency,
     category: &TradeCategory,
@@ -68,7 +68,7 @@ pub fn create_entry(
 
 pub fn create_target(
     trading_vehicle_id: Uuid,
-    quantity: i64,
+    quantity: Decimal,
     price: Decimal,
     currency: &Currency,
     category: &TradeCategory,
